// src/types/generated/ctp.ts
export enum TChannelRole {
  /** Role tells that this channel can be used to track inventory entries.Channels with this role can be treated as warehouses */
  InventorySupply = 'InventorySupply',
  /** Role tells that this channel can be used to track order export activities. */
  OrderExport = 'OrderExport',
  /** Role tells that this channel can be used to track order import activities. */
  OrderImport = 'OrderImport',
  /** This role can be combined with some other roles (e.g. with `InventorySupply`) to represent the fact that this particular channel is the primary/master channel among the channels of the same type. */
  Primary = 'Primary',
  /** Role tells that this channel can be used to expose products to a specific distribution channel. It can be used by the cart to select a product price. */
  ProductDistribution = 'ProductDistribution',
}
// ...
export type TFetchReviewsQuery = {
  __typename?: 'Query';
  reviews: {
    __typename?: 'ReviewQueryResult';
    total: number;
    count: number;
    offset: number;
    results: Array<{
      __typename?: 'Review';
      id: string;
      key: string;
      rating: string,
      roles: Array<TChannelRole>;
      nameAllLocales?: Array<{
        __typename?: 'LocalizedString';
        locale: string;
        value: string;
      }> | null;
    }>;
  };
};